test: Harden Redis empty-array preservation
